requirements.txt
telethon
websockets
pysocks